+++ tchannel/thrift/rw.py
@@ -308,10 +308,9 @@ def register(dispatcher, service, handler=None, method=None):
         )
         assert not function.oneway
 
-        handler = build_handler(function, handler)
         dispatcher.register(
             function.endpoint,
-            handler,
+            build_handler(function, handler),
             ThriftRWSerializer(service._module, function._request_cls),
             ThriftRWSerializer(service._module, function._response_cls),
         )